refactor: more namespace pollution fixing
