+++ src/prelude.rs
@@ -1,0 +1,10 @@
+pub use crate::symbolic_domain::{
+    BinaryIntegerDomain,
+    GrayCodeIntegerDomain,
+    PetriNetIntegerDomain,
+    // todo uncomment one those working
+    // GenericIntegerDomain,
+    // GenericStateSpaceDomain,
+    SymbolicDomain,
+    UnaryIntegerDomain,
+};
