+++ src/symbolic_domain.rs
@@ -1,3 +1,8 @@
+//! A private module which stores the implementation of the traits/structures relevant for
+//! symbolic encoding of logical models.
+//!
+//! TODO:
+//!     In the final library, we should re-export the relevant types from this module here.
 use biodivine_lib_bdd::{
     Bdd, BddPartialValuation, BddVariable, BddVariableSet, BddVariableSetBuilder,
 };
